Differentiate between full bidding that will eventually be handled by TalentMAP, versus what will be available in FSBid; set this in a feature flag
