+++ public/config/config.json
@@ -6,6 +6,7 @@
     "bidding": true,
     "projected_vacancy": true,
     "static_content": true,
-    "notifications": true
+    "notifications": true,
+    "complete_bidding": true
   }
 }
